requirements.txt
langchain-openai
pandas
numpy
seaborn
python-dotenv
langchain_community
langchain_core
streamlit